+++ docs/CHANGELOG.md
@@ -5,3 +5,4 @@ Changelog
 -------------------------
 - Init: Initialize Module "Text editor"
 - Enh #1: Store file version on update content
+- Fix #5: Position of "Edit" Handler